refactor(exchange): 환율 provider를 out port 패키지로 분리
- ExchangeRateProvider와 관련 예외를 application/port/outbound 패키지로 이동
- 기존 layered architecture 안에서 외부 provider 계약의 위치를 명시
- Infrastructure adapter의 application 계약 구조 유지

Refs: #88
