+++ src/main/java/com/lemonpay/exchange/application/port/outbound/ExchangeRateProvider.java
@@ -1,6 +1,7 @@
-package com.lemonpay.exchange.application;
+package com.lemonpay.exchange.application.port.outbound;
 
 import com.lemonpay.common.domain.Currency;
+import com.lemonpay.exchange.application.ExchangeRateSnapshot;
 
 import java.time.LocalDate;
