+++ src/main/java/com/lemonpay/exchange/infrastructure/provider/exchangeapi/ExchangeRateApiAdapter.java
@@ -3,8 +3,8 @@ package com.lemonpay.exchange.infrastructure.provider.exchangeapi;
 import com.lemonpay.common.domain.Currency;
 import com.lemonpay.common.exception.CoreException;
 import com.lemonpay.common.exception.ErrorType;
-import com.lemonpay.exchange.application.ExchangeRateProvider;
-import com.lemonpay.exchange.application.ExchangeRateProviderException;
+import com.lemonpay.exchange.application.port.outbound.ExchangeRateProvider;
+import com.lemonpay.exchange.application.port.outbound.ExchangeRateProviderException;
 import com.lemonpay.exchange.application.ExchangeRateSnapshot;
 import com.lemonpay.exchange.domain.ExchangeRateSource;
 import com.lemonpay.exchange.domain.ExchangeRateType;
